Extract Version and ModuleSpec types into modules of their own
